Fix macro to accept trailing comma
If the `nonempty!` macro was used with a trailing comma:

    let v = nonempty![1,];

then it would fail to compile due to the trailing comma.

The `vec!` macro[[0]] uses `$(,)?` to handle this case:

    ($($x:expr),+ $(,)?)

To fix the `nonempty!` macro, the same is done and a test case is
added to its documentation.

[0]: https://doc.rust-lang.org/src/alloc/macros.rs.html#49

Signed-off-by: Fintan Halpenny <fintan.halpenny@gmail.com>
X-Clacks-Overhead: GNU Terry Pratchett