feat: update bincode optional feature
It was a misnomer to name the feature flag `bincode2` since that is the name of
unmaintained library. The library `bincode` is actively maintained, so the
feature flag is named after it instead.

There is a proper version release v2.0.1 now, which is specified instead. The
derive is cleaned up, and there is a roundtrip test added.