cargo: MSRV and GitHub Actions
The aim of this change is to ensure that `nonempty` maintains its minimal supported
rust version (MSRV). This is declared in the `Cargo.toml`, and is currently `1.79.0`.

However, this version differs depending on which features are used:
- `serialize` requires 1.81.0
- `bincode` requires 1.85.1

The GitHub actions are changed so that we split these checks up into their categories:
- MSRV performs the full check using the 1.79.0 toolchain
- `Serde Build` ensures that the `serialize` feature builds and tests
- `Bincode Build` ensures that the `bincode` feature builds and tests
- `no-std` ensures that no-std builds and tests

While changing the actions, it is important to note that `actions-rs` has been
unmaintained, and so a switch to
[`dtolnay/rust-toolchain`](https://github.com/dtolnay/rust-toolchain) is made,
as it is better maintained by a prominent Rust developer – and the project is
still active.